Update donation info
