+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
       android:resizeableActivity="true"
       android:theme="@style/Theme.EasyGif.Dialog" />
     <activity
-      android:name=".FollowWechatActivity"
+      android:name=".DonateCryptoActivity"
       android:configChanges="mcc|mnc|locale|touchscreen|keyboard|keyboardHidden|navigation|screenLayout|fontScale|uiMode|orientation|density|screenSize|smallestScreenSize"
       android:exported="false"
       android:launchMode="singleTask"
