+++ app/src/main/res/layout/activity_donate.xml
@@ -65,7 +65,7 @@
           android:layout_centerInParent="true"
           android:layout_gravity="center"
           android:layout_marginVertical="12dp"
-          app:srcCompat="@raw/donate_wechat" />
+          app:srcCompat="@raw/donate_buymeacoffee" />
 
         <com.google.android.material.textview.MaterialTextView
           android:layout_width="match_parent"
@@ -119,7 +119,7 @@
           android:layout_height="wrap_content"
           android:layout_marginHorizontal="8dp"
           android:layout_weight="1"
-          android:text="@string/save_donate_qrcode_open_wechat_scan"
+          android:text="@string/save_donate_qrcode_buy_me_a_coffee"
           android:textColor="?attr/colorOnSurface" />
       </androidx.appcompat.widget.LinearLayoutCompat>
     </androidx.appcompat.widget.LinearLayoutCompat>
