+++ app/src/main/res/values/strings.xml
@@ -21,7 +21,7 @@
   <string name="gif_duration_s">GIF duration: %s seconds</string>
   <string name="effect_needs_to_be_viewed_after_exporting">ℹ️ The effects of %s needs to be viewed after exporting</string>
   <string name="beta_info">The current version is for testing purposes only and may contain bugs. \nThe internal test will end in %d days. After the end, please download the official version from my WeChat account.</string>
-  <string name="developer_name_email">Developer: 沈科光 (tasy5kg@qq.com)</string>
+  <string name="developer_name_email">Developer: Tuan Nguyen (tuancoltech@gmail.com)</string>
   <string name="files">Files</string>
   <string name="photos">Photos</string>
   <string name="float_window">Window</string>
@@ -32,7 +32,7 @@
   <string name="rotate">Rotate</string>
   <string name="color">Color</string>
   <string name="input_text_here">Input text here</string>
-  <string name="sorry_easygif_error">Sorry, there was an error while running EasyGIF. \nI hope you can send the problem log to the developer\'s WeChat public account [光光可爱小仙女] so that this error can be fixed.</string>
+  <string name="sorry_easygif_error">Sorry, there was an error while running EasyGIF. \nI hope you can send the problem log to the developer\'s email [tuancoltech@gmail.com] so that this error can be fixed.</string>
   <string name="error_log">Error log</string>
   <string name="copy_to_clipboard">Copy to clipboard</string>
   <string name="follow_wechat_to_feedback">Follow my WeChat public account [光光可爱小仙女] to provide feedback</string>
@@ -44,10 +44,10 @@
   <string name="if_you_like_easygif_please_donate">If you like EasyGIF, please donate to me to support me in developing more useful Apps!</string>
   <string name="thanks_to_the_following_donated_users">Thanks to the following donated users:</string>
   <string name="you_can_write_your_nickname">You can write your nickname in the payment remark so that your nickname will be displayed in this list when this software is updated.</string>
-  <string name="save_donate_qrcode_open_wechat_scan">Save the QR code</string>
+  <string name="save_donate_qrcode_buy_me_a_coffee">Save the QR code</string>
   <string name="_eula_">"User Agreement"</string>
   <string name="eula_line_1">Welcome to use EasyGIF!</string>
-  <string name="eula_line_2">EasyGIF is an open source, ad-free, free-to-use GIF processing software developed by 沈科光.</string>
+  <string name="eula_line_2">EasyGIF is an open source, ad-free, free-to-use GIF processing software developed by Tuan Nguyen.</string>
   <string name="eula_line_3">Disclaimer: This software is provided as is without warranty of any kind. The author is not responsible for any damages, claims or other liability arising out of the use or otherwise of the software.</string>
   <string name="_privacy_policy_">"Privacy Policy"</string>
   <string name="privacy_policy_line_1">This software does not connect to the Internet, does not require any permissions, and does not collect any personal information.</string>
@@ -56,10 +56,9 @@
   <string name="gif_saved_to_gallery">GIF saved to gallery</string>
   <string name="video_saved_to_gallery">Video saved to gallery</string>
   <string name="delete">Delete</string>
-  <string name="follow_wechat">Follow my WeChat public account</string>
-  <string name="follow_my_wechat">Follow my WeChat public account</string>
-  <string name="follow_wechat_line_1">Follow my WeChat public account [光光可爱小仙女] to download and update EasyGIF more conveniently. \nIf you have any questions or suggestions, you can also send me a message on the official account. I will read every message.</string>
-  <string name="save_qrcode_open_wechat_scan">Save the QR code</string>
+  <string name="donate_with_crypto">Donate with crypto</string>
+  <string name="donate_with_crypto_line_1">If you find this app useful, you may consider supporting the author with a crypto donation to encourage continued development of valuable work.\nPlease send only assets supported by Ethereum / ERC-20 token to the address shown below. Transfers on unsupported networks may be unrecoverable.</string>
+  <string name="save_qrcode">Save the QR code</string>
   <string name="gif_split">GIF Split</string>
   <string name="save_this_frame">Save this frame</string>
   <string name="previous_frame">Previous frame</string>
@@ -128,18 +127,16 @@
   <string name="exporting_gif_">Exporting GIF…</string>
   <string name="unable_to_load_gif">Unable to load GIF</string>
   <string name="gif_to_video_conversion_failed">GIF to video conversion failed</string>
-  <string name="wechat_follow_qrcode_saved_please_scan_first_image">My Wechat account QR code has been saved, please scan the first picture in your gallery</string>
-  <string name="wechat_donate_qrcode_saved_please_scan_first_image">Wechat donation QR code has been saved, please scan the first picture in your gallery</string>
-  <string name="wechat_donate_qrcode_saved_wechat_launch_failed">Wechat donation QR code has been saved</string>
-  <string name="wechat_follow_qrcode_saved_wechat_launch_failed">My Wechat account QR code has been saved</string>
+  <string name="donation_erc20_qrcode_saved_please_scan_first_image">ERC20 donation QR code has been saved, please scan the first picture in your gallery</string>
+  <string name="donate_qrcode_saved_please_scan_first_image">Donation QR code has been saved, please scan the first picture in your gallery</string>
   <string name="crop_4_3" translatable="false">4:3</string>
   <string name="crop_16_9" translatable="false">16:9</string>
   <string name="qq_group_id" translatable="false">729757909</string>
-  <string name="_donated_users_list" translatable="false">\@小二哥哥 @杨啊楊 @华 @Z*g @张先森 @禽兽养殖二厂</string>
+  <string name="_donated_users_list" translatable="false"></string>
   <string name="github_com_tasy5kg_easygif" translatable="false">https://github.com/tuancoltech/EasyGIF</string>
   <string name="_open_source_project_used_links" translatable="false">github.com/androidx/androidx\ngithub.com/material-components/material-components-android\ngithub.com/bumptech/glide\ngithub.com/CanHub/Android-Image-Cropper\ngithub.com/arthenica/ffmpeg-kit\ngithub.com/Tencent/MMKV\ngithub.com/kohler/gifsicle\ngithub.com/santaevpavel/OutlineSpan</string>
   <string name="eula_line_addition">This software is for learning and communication purposes. It does not access the network and does not provide Internet information services.</string>
-  <string name="_my_wechat_public_account_name" translatable="false">光光可爱小仙女</string>
+  <string name="donation_crypto_supported_chain">Supported chain: Ethereum / ERC-20 tokens</string>
   <string name="buy_me_a_cup_of_coffee">Buy me a cup of coffee</string>
   <string name="import_file_failed_please_try">File read failed, please try switching the file picker in the menu</string>
   <string name="save_as_video">Save as video</string>
